Fix to allow using Smalltalk blocks even after the method containing their definition has returned.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/steptalk/trunk@36420 72102866-910b-0410-8b05-ffd578937521

+++ Languages/Smalltalk/STExecutionContext.h
@@ -33,7 +33,7 @@
     
     STStack            *stack;
     
-    unsigned            instructionPointer;
+    NSUInteger          instructionPointer;
 }
 - initWithStackSize:(unsigned)stackSize;
